refactor: move recover flows to start not build
